pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legacy-mcp"
version = "0.1.0"
description = "Active Directory MCP Server for AI-powered assessment"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pywinrm>=0.4.3",
    "pyyaml>=6.0",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
legacy-mcp = "legacy_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/legacy_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
